split code over files, added data preview function for table
