+++ packages/file-mapper/package.json
@@ -6,6 +6,7 @@
   "version": "1.0.0",
   "dependencies": {
     "@dans-framework/utils": "workspace:*",
+    "@mui/icons-material": "^5.14.3",
     "@mui/material": "^5.14.3",
     "@reduxjs/toolkit": "^1.9.5",
     "i18next": "^23.4.1",
