+++ packages/file-mapper/src/features/fileMapperSlice.ts
@@ -1,6 +1,6 @@
 import { createSlice, type PayloadAction } from "@reduxjs/toolkit";
 import { RootState } from "../redux/store";
-import type { SerializedFile, Mapping, FileError } from "../types";
+import type { SerializedFile, Mapping, FileError, SheetData } from "../types";
 import type { FormConfig } from "@dans-framework/deposit";
 
 const initialState: {
@@ -9,16 +9,16 @@ const initialState: {
   file: SerializedFile | undefined;
   savedMap: string;
   form: FormConfig | undefined;
-  fileCols: string[] | undefined;
   fileError: string | undefined;
+  fileData: SheetData[] | undefined;
 } = {
   activeStep: 0,
   mapping: {},
   file: undefined,
   savedMap: '',
   form: undefined,
-  fileCols: undefined,
   fileError: undefined,
+  fileData: undefined,
 };
 
 export const fileMapperSlice = createSlice({
@@ -43,8 +43,11 @@ export const fileMapperSlice = createSlice({
     saveData: (state, action) => {
       state.form = action.payload;
     },
-    setFileCols: (state, action: PayloadAction<string[]>) => {
-      state.fileCols = action.payload;
+    setFileData: (state, action: PayloadAction<SheetData[]>) => {
+      state.fileData = action.payload;
+    },
+    resetFileData: (state) => {
+      state.fileData = initialState.fileData;
     },
     setFileError: (state, action: PayloadAction<FileError | undefined>) => {
       state.fileError = action.payload;
@@ -52,20 +55,28 @@ export const fileMapperSlice = createSlice({
     resetFileError: (state) => {
       state.fileError = initialState.fileError;
     },
-    resetFileCols: (state) => {
-      state.fileCols = initialState.fileCols;
-    }
   },
 });
 
-export const { setActiveStep, setMapping, resetMapping, setSavedMap, setFile, saveData, setFileCols, resetFileCols, setFileError, resetFileError } = fileMapperSlice.actions;
+export const { 
+  setActiveStep, 
+  setMapping, 
+  resetMapping, 
+  setSavedMap, 
+  setFile, 
+  saveData, 
+  setFileData,
+  resetFileData, 
+  setFileError, 
+  resetFileError,
+} = fileMapperSlice.actions;
 
 export const getActiveStep = (state: RootState) => state.fileMapper.activeStep;
 export const getMapping = (state: RootState) => state.fileMapper.mapping;
 export const getSavedMap = (state: RootState) => state.fileMapper.savedMap;
 export const getFile = (state: RootState) => state.fileMapper.file;
 export const getForm = (state: RootState) => state.fileMapper.form;
-export const getFileCols = (state: RootState) => state.fileMapper.fileCols;
+export const getFileData = (state: RootState) => state.fileMapper.fileData;
 export const getFileError = (state: RootState) => state.fileMapper.fileError;
 
 export default fileMapperSlice.reducer;
